// src/api/model/sourceDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vector
 * The Vector API
 * OpenAPI spec version: 1.0
 */
import type { SourceDtoDeletedAt } from './sourceDtoDeletedAt';
import type { SourceDtoDeletedBy } from './sourceDtoDeletedBy';

export interface SourceDto {
  createdAt: string;
  createdBy: number;
  updatedAt: string;
  updatedBy: number;
  /** @nullable */
  deletedAt: SourceDtoDeletedAt;
  /** @nullable */
  deletedBy: SourceDtoDeletedBy;
  id: number;
  name: string;
  workspaceId: number;
}